// src/components/units/carPickKey/before/CarPickKey.before.types.ts
import { ReactNativeFile } from "apollo-upload-client";
import { Dispatch, SetStateAction } from "react";

export interface ICarPickKeyBeforeUIProps {
   imageFiles: ReactNativeFile[];
   setImageFiles: Dispatch<SetStateAction<ReactNativeFile[]>>;
   imageUris: string[];
   setImageUris: Dispatch<SetStateAction<string[]>>;
   onPressToCarPickKey: () => void;
}